restaurant-backend/requirements.txt

fastapi
uvicorn
gunicorn
pydantic
supabase
python-dotenv
google-generativeai